requirements.txt
python_telegram_bot==11.1.0
vk_api==11.8.0
dialogflow==1.0.0
python-dotenv==0.14.0
telegram==0.0.1
